[REF] website_sale: split product/template model in 2 files
