+++ addons/website_sale/models/__init__.py
@@ -4,19 +4,21 @@
 from . import account_move
 from . import base_unit
 from . import crm_team
+from . import digest
 from . import ir_http
 from . import mail_compose_message
 from . import payment
+from . import product_attribute
 from . import product_image
-from . import product
+from . import product_misc
+from . import product_product
+from . import product_template
+from . import res_company
+from . import res_config_settings
 from . import res_country
 from . import res_partner
 from . import sale_order
 from . import website
-from . import res_config_settings
-from . import digest
-from . import res_company
-from . import product_attribute
 from . import website_page
-from . import website_visitor
 from . import website_snippet_filter
+from . import website_visitor
