+++ addons/website_sale/models/product_misc.py
@@ -1,0 +1,178 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from odoo import api, fields, models, tools, _
+from odoo.exceptions import ValidationError, UserError
+from odoo.addons.website.models import ir_http
+from odoo.tools.translate import html_translate
+
+
+class ProductRibbon(models.Model):
+    _name = "product.ribbon"
+    _description = 'Product ribbon'
+
+    def name_get(self):
+        return [(ribbon.id, '%s (#%d)' % (tools.html2plaintext(ribbon.html), ribbon.id)) for ribbon in self]
+
+    html = fields.Html(string='Ribbon html', required=True, translate=True, sanitize=False)
+    bg_color = fields.Char(string='Ribbon background color', required=False)
+    text_color = fields.Char(string='Ribbon text color', required=False)
+    html_class = fields.Char(string='Ribbon class', required=True, default='')
+
+
+class ProductPricelist(models.Model):
+    _inherit = "product.pricelist"
+
+    def _default_website(self):
+        """ Find the first company's website, if there is one. """
+        company_id = self.env.company.id
+
+        if self._context.get('default_company_id'):
+            company_id = self._context.get('default_company_id')
+
+        domain = [('company_id', '=', company_id)]
+        return self.env['website'].search(domain, limit=1)
+
+    website_id = fields.Many2one('website', string="Website", ondelete='restrict', default=_default_website, domain="[('company_id', '=?', company_id)]")
+    code = fields.Char(string='E-commerce Promotional Code', groups="base.group_user")
+    selectable = fields.Boolean(help="Allow the end user to choose this price list")
+
+    def clear_cache(self):
+        # website._get_pl_partner_order() is cached to avoid to recompute at each request the
+        # list of available pricelists. So, we need to invalidate the cache when
+        # we change the config of website price list to force to recompute.
+        website = self.env['website']
+        website._get_pl_partner_order.clear_cache(website)
+
+    @api.model
+    def create(self, data):
+        if data.get('company_id') and not data.get('website_id'):
+            # l10n modules install will change the company currency, creating a
+            # pricelist for that currency. Do not use user's company in that
+            # case as module install are done with OdooBot (company 1)
+            self = self.with_context(default_company_id=data['company_id'])
+        res = super(ProductPricelist, self).create(data)
+        self.clear_cache()
+        return res
+
+    def write(self, data):
+        res = super(ProductPricelist, self).write(data)
+        if data.keys() & {'code', 'active', 'website_id', 'selectable', 'company_id'}:
+            self._check_website_pricelist()
+        self.clear_cache()
+        return res
+
+    def unlink(self):
+        res = super(ProductPricelist, self).unlink()
+        self._check_website_pricelist()
+        self.clear_cache()
+        return res
+
+    def _get_partner_pricelist_multi_search_domain_hook(self, company_id):
+        domain = super(ProductPricelist, self)._get_partner_pricelist_multi_search_domain_hook(company_id)
+        website = ir_http.get_request_website()
+        if website:
+            domain += self._get_website_pricelists_domain(website.id)
+        return domain
+
+    def _get_partner_pricelist_multi_filter_hook(self):
+        res = super(ProductPricelist, self)._get_partner_pricelist_multi_filter_hook()
+        website = ir_http.get_request_website()
+        if website:
+            res = res.filtered(lambda pl: pl._is_available_on_website(website.id))
+        return res
+
+    def _check_website_pricelist(self):
+        for website in self.env['website'].search([]):
+            if not website.pricelist_ids:
+                raise UserError(_("With this action, '%s' website would not have any pricelist available.") % (website.name))
+
+    def _is_available_on_website(self, website_id):
+        """ To be able to be used on a website, a pricelist should either:
+        - Have its `website_id` set to current website (specific pricelist).
+        - Have no `website_id` set and should be `selectable` (generic pricelist)
+          or should have a `code` (generic promotion).
+        - Have no `company_id` or a `company_id` matching its website one.
+
+        Note: A pricelist without a website_id, not selectable and without a
+              code is a backend pricelist.
+
+        Change in this method should be reflected in `_get_website_pricelists_domain`.
+        """
+        self.ensure_one()
+        if self.company_id and self.company_id != self.env["website"].browse(website_id).company_id:
+            return False
+        return self.website_id.id == website_id or (not self.website_id and (self.selectable or self.sudo().code))
+
+    def _get_website_pricelists_domain(self, website_id):
+        ''' Check above `_is_available_on_website` for explanation.
+        Change in this method should be reflected in `_is_available_on_website`.
+        '''
+        company_id = self.env["website"].browse(website_id).company_id.id
+        return [
+            '&', ('company_id', 'in', [False, company_id]),
+            '|', ('website_id', '=', website_id),
+            '&', ('website_id', '=', False),
+            '|', ('selectable', '=', True), ('code', '!=', False),
+        ]
+
+    def _get_partner_pricelist_multi(self, partner_ids, company_id=None):
+        ''' If `property_product_pricelist` is read from website, we should use
+            the website's company and not the user's one.
+            Passing a `company_id` to super will avoid using the current user's
+            company.
+        '''
+        website = ir_http.get_request_website()
+        if not company_id and website:
+            company_id = website.company_id.id
+        return super(ProductPricelist, self)._get_partner_pricelist_multi(partner_ids, company_id)
+
+    @api.constrains('company_id', 'website_id')
+    def _check_websites_in_company(self):
+        '''Prevent misconfiguration multi-website/multi-companies.
+           If the record has a company, the website should be from that company.
+        '''
+        for record in self.filtered(lambda pl: pl.website_id and pl.company_id):
+            if record.website_id.company_id != record.company_id:
+                raise ValidationError(_("""Only the company's websites are allowed.\nLeave the Company field empty or select a website from that company."""))
+
+
+class ProductPublicCategory(models.Model):
+    _name = "product.public.category"
+    _inherit = ["website.seo.metadata", "website.multi.mixin", 'image.mixin']
+    _description = "Website Product Category"
+    _parent_store = True
+    _order = "sequence, name, id"
+
+    def _default_sequence(self):
+        cat = self.search([], limit=1, order="sequence DESC")
+        if cat:
+            return cat.sequence + 5
+        return 10000
+
+    name = fields.Char(required=True, translate=True)
+    parent_id = fields.Many2one('product.public.category', string='Parent Category', index=True, ondelete="cascade")
+    parent_path = fields.Char(index=True)
+    child_id = fields.One2many('product.public.category', 'parent_id', string='Children Categories')
+    parents_and_self = fields.Many2many('product.public.category', compute='_compute_parents_and_self')
+    sequence = fields.Integer(help="Gives the sequence order when displaying a list of product categories.", index=True, default=_default_sequence)
+    website_description = fields.Html('Category Description', sanitize_attributes=False, translate=html_translate, sanitize_form=False)
+    product_tmpl_ids = fields.Many2many('product.template', relation='product_public_category_product_template_rel')
+
+    @api.constrains('parent_id')
+    def check_parent_id(self):
+        if not self._check_recursion():
+            raise ValueError(_('Error ! You cannot create recursive categories.'))
+
+    def name_get(self):
+        res = []
+        for category in self:
+            res.append((category.id, " / ".join(category.parents_and_self.mapped('name'))))
+        return res
+
+    def _compute_parents_and_self(self):
+        for category in self:
+            if category.parent_path:
+                category.parents_and_self = self.env['product.public.category'].browse([int(p) for p in category.parent_path.split('/')[:-1]])
+            else:
+                category.parents_and_self = category
